Create_AWS_Artifacts/requirements.txt
toml==0.10.2
python-dotenv==0.20.0